Moved CSS files to styles folder
