+++ components/Footer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Footer.module.css";
+import "../styles/Footer.module.css";
 
 function Footer() {
   return (
